# backend/gestion-evento-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión de Eventos",
        "description": "Documentación de la API para el proyecto de gestión de eventos",
        "contact": {
            "email": "soporte@ejemplo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/facultades": {
            "get": {
                "tags": [
                    "Facultades"
                ],
                "summary": "Listar facultades",
                "description": "Devuelve todos las facultades disponibles",
                "operationId": "58f91da0e2c917d30548e37ea0a681d1",
                "responses": {
                    "200": {
                        "description": "Lista de facultades",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Facultad"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Facultades"
                ],
                "summary": "Crear una nueva facultad",
                "description": "Crea una facultad con nombre y opcionalmente una foto",
                "operationId": "5408641f6336fb051b4adfd4243920b0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "description": "Nombre de la facultad",
                                        "type": "string",
                                        "example": "Ingeniería"
                                    },
                                    "codigo": {
                                        "description": "Código de la facultad",
                                        "type": "string",
                                        "example": "ING"
                                    },
                                    "foto": {
                                        "description": "Imagen de la facultad (archivo opcional)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "filial_id": {
                                        "description": "ID de la filial asociada",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Facultad creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Facultad"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El campo nombre es obligatorio."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/facultades/{id}": {
            "get": {
                "tags": [
                    "Facultades"
                ],
                "summary": "Obtener una facultad por ID",
                "description": "Devuelve una facultad específica por su ID",
                "operationId": "fdd038f0941a45a982f4946a1be62c7a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la facultad",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Facultad encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Facultad"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Facultad no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Facultad no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Facultades"
                ],
                "summary": "Actualizar una facultad existente",
                "description": "Permite actualizar el nombre de una facultad y opcionalmente su foto",
                "operationId": "7572a6d7c0d64e45c0535cc4085b2cac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la facultad a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "description": "Nombre de la facultad",
                                        "type": "string",
                                        "example": "Ingeniería"
                                    },
                                    "codigo": {
                                        "description": "Código de la facultad",
                                        "type": "string",
                                        "example": "ING"
                                    },
                                    "foto": {
                                        "description": "Imagen de la facultad (archivo opcional)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "filial_id": {
                                        "description": "ID de la filial asociada",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Facultad actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Facultad"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Facultad no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Facultad no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El campo nombre es obligatorio."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Facultades"
                ],
                "summary": "Eliminar una facultad",
                "description": "Elimina una facultad por su ID",
                "operationId": "10f73686c95d28bd7b1c1805df362b01",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la facultad a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Facultad eliminada correctamente (sin contenido)"
                    },
                    "404": {
                        "description": "Facultad no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Facultad no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Facultad": {
                "title": "Facultad",
                "description": "Facultad del sistema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Example"
                    },
                    "codigo": {
                        "type": "string",
                        "example": "Example"
                    },
                    "foto": {
                        "type": "string",
                        "format": "url",
                        "example": "http://localhost/storage/roles/admin.png"
                    },
                    "filialId": {
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Introduce el token de acceso obtenido al iniciar sesión con Laravel Passport",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}